Fixes for regressions found in testing.

+++ src/com/palmergames/bukkit/towny/command/TownCommand.java
@@ -2944,24 +2944,20 @@ public class TownCommand extends BaseCommand implements CommandExecutor {
 		town.setSpawn(spawn);
 
 		if (world.isUsingPlotManagementRevert()) {
-			if (TownyRegenAPI.getRegenQueueList().contains(townBlock.getWorldCoord())) {
+			PlotBlockData plotChunk = TownyRegenAPI.getPlotChunk(townBlock);
+			if (plotChunk != null && TownyRegenAPI.getRegenQueueList().contains(townBlock.getWorldCoord())) {
 				// This plot is in the regeneration queue.
-				PlotBlockData plotChunk = TownyRegenAPI.getPlotChunk(townBlock);
-				if (plotChunk != null) {
-
-					TownyRegenAPI.deletePlotChunk(plotChunk); // just claimed so stop regeneration.
-	
-				} else {
-	
-					plotChunk = new PlotBlockData(townBlock); // Not regenerating so create a new snapshot.
-					plotChunk.initialize();
-	
-				}
-				TownyRegenAPI.addPlotChunkSnapshot(plotChunk); // Save a snapshot.
-				plotChunk = null;
+				TownyRegenAPI.deletePlotChunk(plotChunk); // just claimed so stop regeneration.
 				// Remove the WorldCoord from the regenqueue.
 				TownyRegenAPI.removeFromRegenQueueList(townBlock.getWorldCoord());
+			} else {
+
+				plotChunk = new PlotBlockData(townBlock); // Not regenerating so create a new snapshot.
+				plotChunk.initialize();
+
 			}
+			TownyRegenAPI.addPlotChunkSnapshot(plotChunk); // Save a snapshot.
+			plotChunk = null;
 		}
 		if (TownyEconomyHandler.isActive()) {
 			TownyMessaging.sendDebugMsg("Creating new Town account: " + TownySettings.getTownAccountPrefix() + name);
